Adicionando novas funcionalidades e adicionando mais seguranca para evitar acessos indevidos por links na autorizados

+++ src/main/java/com/mballem/curso/security/web/controller/EspecialidadeController.java
@@ -50,8 +50,13 @@ public class EspecialidadeController {
 
 	@GetMapping("/excluir/{id}")
 	public String excluir(@PathVariable("id") Long id, RedirectAttributes attr) {
-		service.remover(id);
-		attr.addFlashAttribute("Sucesso", "Operação realizada com sucesso.");
+		
+		if(service.existeEspecialidadeEmAgendamento(id)) {
+			attr.addFlashAttribute("falha", "Existem consultas agendados para essa especialidade, exclusão negada");
+		}else{
+			service.remover(id);
+			attr.addFlashAttribute("sucesso", "Operação realizada com sucesso.");
+		}
 		return "redirect:/especialidades";
 	}
